Instance satisfy refactored
