+++ src/main/kotlin/com/cognifide/gradle/aem/common/instance/AbstractInstance.kt
@@ -64,6 +64,8 @@ abstract class AbstractInstance(@Transient @JsonIgnore protected val aem: AemExt
 
     fun check(options: CheckAction.() -> Unit) = manager.check(this, options)
 
+    fun provision() = manager.provisioner.provision(this)
+
     override fun equals(other: Any?): Boolean {
         if (this === other) return true
         if (javaClass != other?.javaClass) return false
