+++ src/main/kotlin/com/cognifide/gradle/aem/common/instance/provision/Action.kt
@@ -1,0 +1,3 @@
+package com.cognifide.gradle.aem.common.instance.provision
+
+class Action(val step: InstanceStep, val status: Status)
