+++ src/main/kotlin/com/cognifide/gradle/aem/common/instance/provision/Status.kt
@@ -1,0 +1,7 @@
+package com.cognifide.gradle.aem.common.instance.provision
+
+enum class Status {
+    ENDED,
+    FAILED,
+    SKIPPED
+}
