+++ src/main/kotlin/com/cognifide/gradle/aem/common/instance/satisfy/PackageAction.kt
@@ -1,0 +1,6 @@
+package com.cognifide.gradle.aem.common.instance.satisfy
+
+import com.cognifide.gradle.aem.common.instance.Instance
+import java.io.File
+
+class PackageAction(val pkg: File, val instance: Instance)
