+++ src/main/kotlin/com/cognifide/gradle/aem/common/instance/satisfy/PackageGroup.kt
@@ -1,4 +1,4 @@
-package com.cognifide.gradle.aem.instance.satisfy
+package com.cognifide.gradle.aem.common.instance.satisfy
 
 import com.cognifide.gradle.aem.common.instance.Instance
 import com.cognifide.gradle.aem.common.instance.InstanceSync
