+++ src/main/kotlin/com/cognifide/gradle/aem/common/instance/satisfy/PackageResolver.kt
@@ -1,4 +1,4 @@
-package com.cognifide.gradle.aem.instance.satisfy
+package com.cognifide.gradle.aem.common.instance.satisfy
 
 import com.cognifide.gradle.aem.AemExtension
 import com.cognifide.gradle.aem.common.bundle.BundleFile
