+++ src/main/kotlin/com/cognifide/gradle/aem/instance/LocalInstancePlugin.kt
@@ -1,8 +1,8 @@
 package com.cognifide.gradle.aem.instance
 
-import com.cognifide.gradle.aem.instance.provision.InstanceProvision
+import com.cognifide.gradle.aem.instance.tasks.InstanceProvision
 import com.cognifide.gradle.aem.instance.rcp.InstanceRcp
-import com.cognifide.gradle.aem.instance.satisfy.InstanceSatisfy
+import com.cognifide.gradle.aem.instance.tasks.InstanceSatisfy
 import com.cognifide.gradle.aem.instance.tail.InstanceTail
 import com.cognifide.gradle.aem.instance.tasks.*
 import com.cognifide.gradle.common.CommonDefaultPlugin
