--- src/main/kotlin/com/cognifide/gradle/aem/instance/provision/Action.kt
@@ -1,3 +1,0 @@
-package com.cognifide.gradle.aem.instance.provision
-
-class Action(val step: InstanceStep, val status: Status)
