--- src/main/kotlin/com/cognifide/gradle/aem/instance/provision/Status.kt
@@ -1,7 +1,0 @@
-package com.cognifide.gradle.aem.instance.provision
-
-enum class Status {
-    ENDED,
-    FAILED,
-    SKIPPED
-}
