+++ src/main/kotlin/com/cognifide/gradle/aem/instance/tasks/InstanceResolve.kt
@@ -1,19 +1,18 @@
 package com.cognifide.gradle.aem.instance.tasks
 
-import com.cognifide.gradle.aem.common.tasks.LocalInstanceTask
-import com.cognifide.gradle.aem.instance.satisfy.InstanceSatisfy
+import com.cognifide.gradle.aem.AemDefaultTask
 import org.gradle.api.tasks.TaskAction
 
-open class InstanceResolve : LocalInstanceTask() {
-
-    init {
-        description = "Resolves instance files from remote sources before running other tasks"
-    }
+open class InstanceResolve : AemDefaultTask() {
 
     @TaskAction
     fun resolve() {
-        common.tasks.get<InstanceSatisfy>(InstanceSatisfy.NAME).resolvePackages() // more light at first
-        localInstanceManager.resolveSourceFiles()
+        aem.instanceManager.resolveFiles()
+        aem.localInstanceManager.resolveFiles()
+    }
+
+    init {
+        description = "Resolves instance files from remote sources before running other tasks"
     }
 
     companion object {
