+++ src/test/kotlin/com/cognifide/gradle/aem/instance/LocalInstancePluginTest.kt
@@ -3,7 +3,7 @@ package com.cognifide.gradle.aem.instance
 import com.cognifide.gradle.aem.AemExtension
 import com.cognifide.gradle.aem.common.utils.using
 import com.cognifide.gradle.aem.instance.rcp.InstanceRcp
-import com.cognifide.gradle.aem.instance.satisfy.InstanceSatisfy
+import com.cognifide.gradle.aem.instance.tasks.InstanceSatisfy
 import com.cognifide.gradle.aem.instance.tasks.InstanceDown
 import com.cognifide.gradle.aem.instance.tasks.InstanceSetup
 import com.cognifide.gradle.aem.instance.tasks.InstanceUp
